[test] Extract generators
