+++ tests/include/generators/foldable-callable.hpp
