+++ tests/include/generators/runnable-combination.hpp
@@ -6,9 +6,9 @@
 
 #pragma once
 
-#include "composable-callable.hpp"
-#include "foldable-callable.hpp"
-#include "productable-container.hpp"
+#include "generators/composable-callable.hpp"
+#include "generators/foldable-callable.hpp"
+#include "generators/productable-container.hpp"
 
 namespace Productable
 {
