+++ tests/include/idempotent.hpp
@@ -57,9 +57,9 @@ private:
     static constexpr CompletableRawContext completable{};
 };
 
-#include "composable-callable.hpp"
-#include "foldable-callable.hpp"
-#include "productable-container.hpp"
+#include "generators/composable-callable.hpp"
+#include "generators/foldable-callable.hpp"
+#include "generators/productable-container.hpp"
 
 template <typename Given>
 struct CompletableRawGivenContext
