+++ tests/src/instanciation/composable-callable.cpp
@@ -4,7 +4,7 @@
  * SPDX - License - Identifier: MIT
  */
 
-#include "composable-callable.hpp"
+#include "generators/composable-callable.hpp"
 
 template <size_t, size_t, typename>
 struct InstanciatedComposableCallableImpl;
